pyproject.toml
[project]
name = "elo-system-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "gunicorn>=25.2.0",
    "python-dotenv>=1.2.2",
]
